+++ test/test_espeak.py
@@ -72,111 +72,6 @@ def test_arabic():
     out = backend.phonemize(text, sep, False)
     assert out == expected
 
-
-@pytest.fixture
-def langswitch_text():
-    return [
-        "j'aime l'anglais",
-        "j'aime le football",
-        "football",
-        "surtout le real madrid",
-        "n'utilise pas google"]
-
-
-@pytest.mark.skipif(
-    not EspeakBackend.is_espeak_ng(),
-    reason='language switch only exists for espeak-ng')
-@pytest.mark.parametrize('njobs', [1, 3])
-def test_language_switch_keep_flags(caplog, langswitch_text, njobs):
-    backend = EspeakBackend('fr-fr', language_switch='keep-flags')
-    out = backend.phonemize(
-        langswitch_text, separator=Separator(), strip=True, njobs=njobs)
-    assert out == [
-        'ʒɛm lɑ̃ɡlɛ',
-        'ʒɛm lə (en)fʊtbɔːl(fr)',
-        '(en)fʊtbɔːl(fr)',
-        'syʁtu lə (en)ɹiəl(fr) madʁid',
-        'nytiliz pa (en)ɡuːɡəl(fr)']
-
-    messages = [msg[2] for msg in caplog.record_tuples]
-    assert (
-        '4 utterances containing language switches on lines 2, 3, 4, 5'
-        in messages)
-    assert (
-        'language switch flags have been kept (applying "keep-flags" policy)'
-        in messages)
-
-
-@pytest.mark.skipif(
-    not EspeakBackend.is_espeak_ng(),
-    reason='language switch only exists for espeak-ng')
-@pytest.mark.parametrize('njobs', [1, 3])
-def test_language_switch_default(caplog, langswitch_text, njobs):
-    # default behavior is to keep the flags
-    backend = EspeakBackend('fr-fr')
-    out = backend.phonemize(
-        langswitch_text, separator=Separator(), strip=True, njobs=njobs)
-    assert out == [
-        'ʒɛm lɑ̃ɡlɛ',
-        'ʒɛm lə (en)fʊtbɔːl(fr)',
-        '(en)fʊtbɔːl(fr)',
-        'syʁtu lə (en)ɹiəl(fr) madʁid',
-        'nytiliz pa (en)ɡuːɡəl(fr)']
-
-    messages = [msg[2] for msg in caplog.record_tuples]
-    assert (
-        '4 utterances containing language switches on lines 2, 3, 4, 5'
-        in messages)
-    assert (
-        'language switch flags have been kept (applying "keep-flags" policy)'
-        in messages)
-
-
-@pytest.mark.skipif(
-    not EspeakBackend.is_espeak_ng(),
-    reason='language switch only exists for espeak-ng')
-@pytest.mark.parametrize('njobs', [1, 3])
-def test_language_switch_remove_flags(caplog, langswitch_text, njobs):
-    backend = EspeakBackend('fr-fr', language_switch='remove-flags')
-    out = backend.phonemize(
-        langswitch_text, separator=Separator(), strip=True, njobs=njobs)
-    assert out == [
-        'ʒɛm lɑ̃ɡlɛ',
-        'ʒɛm lə fʊtbɔːl',
-        'fʊtbɔːl',
-        'syʁtu lə ɹiəl madʁid',
-        'nytiliz pa ɡuːɡəl']
-
-    messages = [msg[2] for msg in caplog.record_tuples]
-    assert (
-        '4 utterances containing language switches on lines 2, 3, 4, 5'
-        in messages)
-    assert (
-        'language switch flags have been removed '
-        '(applying "remove-flags" policy)'
-        in messages)
-
-
-@pytest.mark.skipif(
-    not EspeakBackend.is_espeak_ng(),
-    reason='language switch only exists for espeak-ng')
-@pytest.mark.parametrize('njobs', [1, 3])
-def test_language_switch_remove_utterance(caplog, langswitch_text, njobs):
-    backend = EspeakBackend('fr-fr', language_switch='remove-utterance')
-    out = backend.phonemize(
-        langswitch_text, separator=Separator(), strip=True, njobs=njobs)
-    assert out == ['ʒɛm lɑ̃ɡlɛ']
-
-    messages = [msg[2] for msg in caplog.record_tuples]
-    assert (
-        'removed 4 utterances containing language switches '
-        '(applying "remove-utterance" policy)'
-        in messages)
-
-    with pytest.raises(RuntimeError):
-        backend = EspeakBackend('fr-fr', language_switch='foo')
-
-
 @pytest.mark.parametrize(
     'text, strip, sep',
     ((t, s, u) for t in [
